reorganize api sort object
