+++ packages/api/src/types.ts
@@ -1,0 +1,5 @@
+import API from './index'
+
+export type hash = string
+
+export type IApi = API
